Spawn dead-letter forwarder process per quorum queue leader
That forwarder subscribes to the quorum queue internal discards queue.

Do not use channels (or even worse AMQP client) since they are protocol specific.

The dead-letter forwarder will need to maintain its own state (like a lightweight channel
in the sense that it needs to keep track of consumer acks and publisher confirms).