rabbit_fifo: refactor decorator notifications
So that they are only evaluated after each Ra "batch" using
the handle aux `eval` message. This results in fewer decorator
evaluations during high throughput. Also fix bug where recreating
a consumer that was previously cancelled but still had pending messages
resulted in the consumer being stuck in cancelled status.

This in turn introduced a bug in single active consumer which will be fixed
in the next commit.