Delete from unsettled target queues when target queue is deleted
Before this commit, messages were stuck in 'pendings' when a target
quorum queue got deleted and thereafter recreated with the same name.

The dlx worker relies on the rabbit_fifo_client to redeliver messages
once the dlx worker has routed to a target quroum queue.

Therefore, we need to delete a target quorum queue from the list of
unsettled target queues when it is deleted.
This is a O(N) operation. However, this should be fine, given that
we do not expect target queues to be deleted often in combintation
with massive fanouts for dead-lettered messages.

(The channel does something similar by removing sequence numbers from
 rabbit_confirms when a target queue is deleted.)

Furthermore, if the dead-letter exchange is the default exchange,
rabbit_router:route/2 will always route to a target queue even if it
does not exist. Therefore, we also check whether the target queues
the dlx worker routes to exist before storing them in the 'unsettled'
list of queues.