WIP Use ets:select/2 to fetch necessary queue record fields
Fetch only the queue record fields that are necessary for delivering a
message.

We call ets:select/2 once with a list of specs for each destination queue.
This method takes 4 minutes with 100k queues.
The CPU flame graph shows all the time is spent in db_prog_match. Once
CPU is busy 100% of the time.
The table key seems to be bound.

The alternative approach to call deliver_to_queues in batches of 1,000
does not result in lower memory usage.

Calling erlang:garbage_collect/0 after each batch does result in lower
memory usage.