Requeue messages if link detached
"A session endpoint can choose to unmap its output handle for a link. In this case, the endpoint MUST
send a detach frame to inform the remote peer that the handle is no longer attached to the link endpoint.
If both endpoints do this, the link MAY return to a fully detached state. Note that in this case the
link endpoints MAY still indirectly communicate via the session, as there could still be active deliveries
on the link referenced via delivery-id." [2.6.4]

"The disposition performative MAY refer to deliveries on links that are no longer attached. As long as
the links have not been closed or detached with an error then the deliveries are still "live" and the
updated state MUST be applied." [2.7.6]

Although the spec allows to settle delivery IDs on detached links, RabbitMQ does not respect the 'closed'
field of the DETACH frame and therefore handles every DETACH frame as closed. Since the link is closed,
we expect every outstanding delivery to be requeued.

In addition to consumer cancellation, detaching a link therefore causes in flight deliveries to be requeued.

Note that this behaviour is different from merely consumer cancellation in AMQP legacy:
"After a consumer is cancelled there will be no future deliveries dispatched to it. Note that there can
still be "in flight" deliveries dispatched previously. Cancelling a consumer will neither discard nor requeue them."
[https://www.rabbitmq.com/consumers.html#unsubscribing]