Set x-* annotations as MQTT User Property
as done by mc_amqpl:protocol_state/2 and
mc_amqp:protocol_state/2

We do not include that commit in 3.13 for now as there is no real use
case for it:
Quorum queue client sets x-delivery-count and
stream queue client sets x-stream-offset.
The MQTT plugin does not consume directly from streams, hence there is
no need to set x-stream-offset.
Whether the message is a redelivery from quorum queues is encoded
already in the DUP flag of the MQTT PUBLISH packet.

Also, if source and destination protocols are both MQTT, we must obey:
> The Server MUST send all User Properties unaltered in a PUBLISH packet when forwarding the Application Message to a Client.

We do know if the source protocol was MQTT in mc_mqtt:protoco_state/2 by
checking whether `Topic =/= undefined`.

However, given we don't have a good use case in 3.13, let's not add x-*
annotations as user property.

We park this commit on a branch in case we need it in the future, e.g.
for https://github.com/rabbitmq/rabbitmq-server/issues/10050